refactor(LagDetect): New lag detect logic.
Hopefully it's more reliably, and not just more trigger happy.
But it will somewhat help, as it now freezes objects immediately,
instead of waiting for `phys_timescale 0` to take effect.

We should look into freezing penetrating props before this point
through. This will even allow us to see the owner of those
props, and tell everyone about that son of a bitch.